+++ excel_ingest/__init__.py
@@ -32,7 +32,7 @@ Quick start::
 from excel_ingest.framework import ExcelIngestFramework, IngestResult
 from excel_ingest.validation import FileValidationResult, ValidationStatus
 from excel_ingest.structure import FileStructureMetadata, FileProcessingConfig, FileStatus
-from excel_ingest.metadata import MetadataExtractionResult, combine_column_records, build_superset_schema
+from excel_ingest.metadata import MetadataExtractionResult, combine_column_records, build_superset_schema, COLUMN_RECORD_FIELDS, SIGNATURE_RECORD_FIELDS
 from excel_ingest.loader import LoadResult
 from excel_ingest.mapping import map_to_canonical, CanonicalMapping, MappingStatus, MappingMethod
 
@@ -47,6 +47,8 @@ __all__ = [
     "MetadataExtractionResult",
     "combine_column_records",
     "build_superset_schema",
+    "COLUMN_RECORD_FIELDS",
+    "SIGNATURE_RECORD_FIELDS",
     "LoadResult",
     "map_to_canonical",
     "CanonicalMapping",